+++ tests/test_job_tracker_v2.py
@@ -1023,8 +1023,10 @@ class TestK8sJobTracker:
         prometheus_mock = mock.Mock()
         prometheus_mock.endpoint = "https://prometheus.test/api/v1"
         prometheus_mock.get_cpu_usage.return_value = 2.34 * 3600
+        prometheus_mock.get_billable_cpu_requested.return_value = prometheus_mock.get_cpu_usage.return_value
         prometheus_mock.get_network_received_usage.return_value = 370841160371254.75
         prometheus_mock.get_memory_usage.return_value = 5.678 * 1024 * 1024 * 3600
+        prometheus_mock.get_billable_memory_requested.return_value = prometheus_mock.get_memory_usage.return_value
         prometheus_mock.get_max_executor_memory_usage.return_value = 3.5
 
         return prometheus_mock
@@ -1660,7 +1662,9 @@ class TestK8sStatusGetter:
 
         prometheus_mock = mock.Mock(Prometheus)
         prometheus_mock.get_cpu_usage.return_value = None
+        prometheus_mock.get_billable_cpu_requested.return_value = None
         prometheus_mock.get_memory_usage.return_value = 0.0
+        prometheus_mock.get_billable_memory_requested.return_value = 0.0
         prometheus_mock.endpoint = "https://prometheus.test/api/v1"
 
         k8s_mock = mock.Mock()