+++ CHANGELOG.md
@@ -21,11 +21,13 @@ without compromising stable operations.
 - `run_udf` now supports running CWL scripts. ([#1476](https://github.com/Open-EO/openeo-geopyspark-driver/issues/1476))
 - new `property_filter_adaptations` feature flag (under `load_stac_feature_flags`) for `load_stac` powered collections to inject a property filtering adapter that translates legacy (e.g. OpenSearch) property filters to STAC property filters (eu-cdse/openeo-cdse-infra#803, eu-cdse/openeo-cdse-infra#804)
 - Guard STAC metadata parsing against invalid "item_assets" usage ([Open-EO/openeo-python-client#853](https://github.com/Open-EO/openeo-python-client/issues/853))
-- load_collection through load_stac: support band name aliases/variants (eu-cdse/openeo-cdse-infra#803, eu-cdse/openeo-cdse-infra#804)
-- load_collection through load_stac: new `apply_sentinel2_reflectance_offset` feature flag (eu-cdse/openeo-cdse-infra#803, eu-cdse/openeo-cdse-infra#804)
+- `load_collection` through `load_stac`: support band name aliases/variants (eu-cdse/openeo-cdse-infra#803, eu-cdse/openeo-cdse-infra#804)
+- `load_collection` through `load_stac`: new `apply_sentinel2_reflectance_offset` feature flag (eu-cdse/openeo-cdse-infra#803, eu-cdse/openeo-cdse-infra#804)
 - `StacApiWorkspace`: do not require `export-workspace-enable-merge` ([#1030](https://github.com/Open-EO/openeo-geopyspark-driver/issues/1030))
 - Support latest `pystac` (currently 1.14.3) ([#1496](https://github.com/Open-EO/openeo-geopyspark-driver/issues/1496))
 - Support property filtering in `load_stac.construct_item_collection` in post-dry-run phase ([#1530](https://github.com/Open-EO/openeo-geopyspark-driver/issues/1530))
+- `load_collection` through `load_stac`: fix resolution selection for low-res/non-UTM cases with aliases ([#1539](https://github.com/Open-EO/openeo-geopyspark-driver/issues/1539))
+- `load_stac` harden STAC "band" metadata parsing when "name" field is missing ([Open-EO/openeo-python-client#860](https://github.com/Open-EO/openeo-python-client/issues/860))
 
 
 ## 0.70.0